Move Node-only schema validation out of parser.ts
parser.ts held the ajv/node:fs/path/url schema validation, so the
browser build of buildSceneFromJson dragged those node builtins in and
Vite warned about externalizing node:path/fs/url. Extracted validate()
and its lazy node imports into a new validate.ts, exported from index.ts
(Node/default entry) but not browser.ts. parser.ts is now import-free of
node builtins, and the docs build is warning-free.

CLI, examples, and fuzz tests import validate from the new module (or
the package index, unchanged for the CLI).
