+++ packages/modular-svg-core/src/index.ts
@@ -4,3 +4,4 @@ export * from "./output.ts";
 export * from "./parser.ts";
 export * from "./solver/index.ts";
 export * from "./svg_ast.ts";
+export * from "./validate.ts";
